+++ src/components/view/news/news.ts
@@ -15,11 +15,8 @@ interface New {
 }
 
 class News {
-    protected draw(data: New[]): void {
-        console.log(data);
-        console.log('h1');
+    public draw(data: New[]): void {
         const news = data.length >= 10 ? data.filter((_item, idx) => idx < 10) : data;
-        console.log(news);
         const fragment = document.createDocumentFragment();
         const newsItemTemp = document.querySelector('#newsItemTemp') as HTMLTemplateElement;
